chore: finish moving sets of phrase corrections
